+++ app/models/engine/expressions/identifier_subject.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Engine
+  module Expressions
+    class IdentifierSubject < Identifier
+      def name
+        "_subject"
+      end
+
+      def storage_name
+        "_s_subject"
+      end
+    end
+  end
+end